requirements.txt
asgiref==3.3.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
Django==3.1.5
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-redis==4.12.1
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
google-api-core==1.25.0
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.3.0
Markdown==3.3.3
Pillow==8.1.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pytz==2020.5
redis==3.5.3
requests==2.25.1
rsa==4.7
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
whitenoise==5.2.0
zipp==3.4.0
